Change from Double to Float

+++ src/main/java/noppes/npcs/attribute/IAttributeInstance.java
@@ -5,6 +5,6 @@ package noppes.npcs.attribute;
  */
 public interface IAttributeInstance {
     AttributeDefinition getAttribute();
-    double getValue();
-    void setValue(double value);
+    float getValue();
+    void setValue(float value);
 }
